requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.4.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==20.1.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
